Support for Namespace object
When creating multi-tenant stacks using PaC approach,
the namespace in which ResourceComposition instance is
created is different than the namespace in which the tenant
workload will run. The tenant namespace may not be known
to the Platform Engineering team.

As part of supporting Namespace object, kubediscovery is now
able to go across namespaces if a Namespace object is being
created as part of the tenant setup Helm chart.