Turning caching on; added profiling support
After doing cpu profiling it was discovered that
we were spending maximum amount of time in dynamic client's List
and Get calls. We have introduced caching for these calls where
we are caching the List and Get outputs and using them whenever needed.
Through the caching we were able to reduce the time by 52%.